+++ packages/cognito/src/types.ts
@@ -2,10 +2,10 @@ export type StateItem = {
     State: string;
     CognitoState: string;
     ExpiryTime : number;
-    UseMock: boolean;
 };
 
 export type SessionStateItem = {
+  LocalCode?: string,
   SessionState: string;
   ApigeeCode: string;
   ExpiryTime: number
